# openapi.yaml
openapi: 3.0.3
info:
  title: PaperStock Unified Trading Platform
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /api/trade/orders:
    post:
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceOrderRequestDto'
      responses:
        '200':
          description: Order response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
  /api/trade/orders/{id}:
    get:
      summary: Get order by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
    delete:
      summary: Cancel order by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Cancelled
  /api/trade/positions:
    get:
      summary: List positions
      responses:
        '200':
          description: Position list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PositionDto'
    post:
      summary: Add or update manual holding (simulator mode)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualHoldingRequestDto'
      responses:
        '200':
          description: Upserted holding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionDto'
  /api/trade/positions/{symbol}:
    delete:
      summary: Delete manual holding by symbol (simulator mode)
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Holding deleted
  /api/account:
    get:
      summary: Get account summary
      responses:
        '200':
          description: Account data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDto'
  /api/market/latest:
    get:
      summary: Get latest market snapshot
      parameters:
        - in: query
          name: symbol
          required: true
          schema:
            type: string
        - in: query
          name: refresh
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Latest market price
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketSnapshotDto'
  /api/market/search:
    get:
      summary: Search stock symbols from configured provider
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: Symbol matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MarketSymbolDto'
  /api/market/config:
    get:
      summary: Get effective market quote configuration
      responses:
        '200':
          description: Active market provider and stale threshold
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketConfigDto'
  /api/strategies:
    get:
      summary: List strategy configurations
      responses:
        '200':
          description: Strategy list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StrategyConfigDto'
    post:
      summary: Create strategy configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StrategyConfigRequestDto'
      responses:
        '200':
          description: Strategy created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyConfigDto'
  /api/strategies/{id}:
    put:
      summary: Update strategy configuration
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StrategyConfigRequestDto'
      responses:
        '200':
          description: Strategy updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyConfigDto'
    delete:
      summary: Delete strategy configuration
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Strategy deleted
  /api/strategies/{id}/pause:
    post:
      summary: Pause strategy
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Strategy paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyConfigDto'
  /api/strategies/{id}/resume:
    post:
      summary: Resume strategy
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Strategy resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyConfigDto'
  /api/strategies/activity:
    get:
      summary: Strategy automation activity feed
      parameters:
        - in: query
          name: symbol
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Strategy activity list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StrategyExecutionDto'
  /api/strategies/report:
    get:
      summary: Aggregated buy/sell report per stock from strategy executions
      responses:
        '200':
          description: Per-symbol buy/sell totals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StrategyTradeReportDto'
  /api/strategies/{id}/executions:
    get:
      summary: Strategy executions by strategy id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Execution history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StrategyExecutionDto'
  /api/strategies/{id}/chart:
    get:
      summary: Chart data for price and trigger lines
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 10
            maximum: 500
      responses:
        '200':
          description: Strategy chart series
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyChartSeriesDto'
components:
  schemas:
    PlaceOrderRequestDto:
      type: object
      required: [symbol, qty, side, type]
      properties:
        symbol:
          type: string
        qty:
          type: number
        side:
          type: string
          enum: [buy, sell]
        type:
          type: string
          enum: [market, limit]
        limitPrice:
          type: number
    OrderDto:
      type: object
      properties:
        id:
          type: string
        symbol:
          type: string
        side:
          type: string
        type:
          type: string
        status:
          type: string
        qty:
          type: number
        filledQty:
          type: number
        averageFillPrice:
          type: number
    TradeDto:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        symbol:
          type: string
        qty:
          type: number
        price:
          type: number
        fee:
          type: number
    PositionDto:
      type: object
      properties:
        symbol:
          type: string
        qty:
          type: number
        averagePrice:
          type: number
    ManualHoldingRequestDto:
      type: object
      required: [symbol, qty, buyPrice]
      properties:
        symbol:
          type: string
        qty:
          type: number
        buyPrice:
          type: number
    AccountDto:
      type: object
      properties:
        mode:
          type: string
        cash:
          type: number
        equity:
          type: number
    MarketSnapshotDto:
      type: object
      properties:
        symbol:
          type: string
        price:
          type: number
        capturedAt:
          type: string
          format: date-time
    StrategyConfigRequestDto:
      type: object
      required:
        - symbol
        - buyDropPercent
        - sellRisePercent
        - buyCashPercent
        - sellPositionPercent
        - maxOrdersPerDay
        - cooldownMinutes
        - active
        - simulatorEnabled
        - alpacaEnabled
      properties:
        symbol:
          type: string
        buyDropPercent:
          type: number
        sellRisePercent:
          type: number
        buyCashPercent:
          type: number
        sellPositionPercent:
          type: number
        maxOrdersPerDay:
          type: integer
        cooldownMinutes:
          type: integer
        active:
          type: boolean
        simulatorEnabled:
          type: boolean
        alpacaEnabled:
          type: boolean
    StrategyConfigDto:
      type: object
      properties:
        id:
          type: string
        symbol:
          type: string
        buyDropPercent:
          type: number
        sellRisePercent:
          type: number
        buyCashPercent:
          type: number
        sellPositionPercent:
          type: number
        maxOrdersPerDay:
          type: integer
        cooldownMinutes:
          type: integer
        active:
          type: boolean
        simulatorEnabled:
          type: boolean
        alpacaEnabled:
          type: boolean
        lastActionAt:
          type: string
          format: date-time
    StrategyExecutionDto:
      type: object
      properties:
        id:
          type: string
        strategyConfigId:
          type: string
        symbol:
          type: string
        side:
          type: string
        triggerPrice:
          type: number
        referencePrice:
          type: number
        orderId:
          type: string
        status:
          type: string
        message:
          type: string
        executedAt:
          type: string
          format: date-time
    StrategyTradeReportDto:
      type: object
      properties:
        symbol:
          type: string
        buyQty:
          type: number
        buyAmount:
          type: number
        buyTrades:
          type: integer
          format: int64
        sellQty:
          type: number
        sellAmount:
          type: number
        sellTrades:
          type: integer
          format: int64
    StrategyChartPointDto:
      type: object
      properties:
        capturedAt:
          type: string
          format: date-time
        price:
          type: number
        buyTriggerPrice:
          type: number
        sellTriggerPrice:
          type: number
    StrategyChartSeriesDto:
      type: object
      properties:
        strategyId:
          type: string
        symbol:
          type: string
        referencePrice:
          type: number
        buyDropPercent:
          type: number
        sellRisePercent:
          type: number
        stale:
          type: boolean
        latestAgeSeconds:
          type: integer
          format: int64
        latestQuoteAt:
          type: string
          format: date-time
        points:
          type: array
          items:
            $ref: '#/components/schemas/StrategyChartPointDto'
    MarketSymbolDto:
      type: object
      properties:
        symbol:
          type: string
        name:
          type: string
        exchange:
          type: string
        type:
          type: string
    MarketConfigDto:
      type: object
      properties:
        provider:
          type: string
        maxStaleSeconds:
          type: integer
          format: int64